Improve enrichment tests
